+++ src/types/socket.d.ts
@@ -11,6 +11,7 @@ export type SocketCommands =
   | "edition"
   | "fabled"
   | "gamePhase"
+  | "annouceWinner"
   | "getGamestate"
   | "gs"
   | "playSound"